+++ src/addons/display_commander/ui/new_ui/hotkeys_tab.hpp
@@ -26,7 +26,6 @@ enum class HotkeyId : int {
     InputBlocking,
     DisplayCommanderUi,
     PerformanceOverlay,
-    Stopwatch,
     WinDown,
     WinUp,
     WinLeft,